+++ Plugins/Creature/Creature.hpp
@@ -9,16 +9,6 @@ using ArgumentStack = NWNXLib::Services::Events::ArgumentStack;
 
 namespace Creature {
 
-namespace CasterType
-{
-enum TYPE
-{
-    None = 0,
-    Arcane = 1,
-    Divine = 2,
-};
-}
-
 class Creature : public NWNXLib::Plugin
 {
 public:
@@ -122,20 +112,13 @@ private:
     CNWSCreature *creature(ArgumentStack& args);
 
 private:
-
-    static uint8_t s_classCasterType[NWNXLib::API::Constants::ClassType::MAX + 1];
-    static uint8_t s_divModClasses[NWNXLib::API::Constants::ClassType::MAX + 1];
-    static uint8_t s_arcModClasses[NWNXLib::API::Constants::ClassType::MAX + 1];
-    static bool s_bUseCasterLevel2da;
-    static bool s_bCasterClassesLoaded;
     static bool s_bAdjustCasterLevel;
+    static bool s_bCasterLevelHooksInitialized;
 
-    static void LoadCasterLevelModifiers();
-    static uint8_t CNWSCreatureStats__GetClassLevel(CNWSCreatureStats* thisPtr, uint8_t nMultiClass, BOOL bUseNegativeLevel);
+    static void InitCasterLevelHooks();
+    static void CNWSCreatureStats__GetClassLevel(bool before, CNWSCreatureStats* thisPtr, uint8_t nMultiClass, BOOL bUseNegativeLevel);
     static void CNWVirtualMachineCommands__ExecuteCommandGetCasterLevel(bool before, CNWVirtualMachineCommands*, int32_t, int32_t);
     static void CNWVirtualMachineCommands__ExecuteCommandResistSpell(bool before, CNWVirtualMachineCommands*, int32_t, int32_t);
-    static void CNWSEffectListHandler__OnApplyDispelAllMagic(bool before, CNWSEffectListHandler*, CNWSObject*, CGameEffect*, BOOL);
-    static void CNWSEffectListHandler__OnApplyDispelBestMagic(bool before, CNWSEffectListHandler*, CNWSObject*, CGameEffect*, BOOL);
     static void CGameEffect__SetCreator(bool before, CGameEffect*, OBJECT_ID);
 
 };